requirements.txt
pyyaml
requests
pyjq